+++ CHANGELOG.md
@@ -60,6 +60,7 @@ All notable changes to this project will be documented in this file.
 - *(unkown)* I forgot but is good to save
 - *(topics)* Added a new solicitation adjustment
 - *(quiz)* Addded some routes for complaining questions
+- *(quiz)* Added a complain for pdf save in redis
 
 ### 🐛 Bug Fixes